style: :truck: Moved screens
Se cambio la estructura de carpetas dentro de screens
