+++ lib/presentation/screens/screens.dart
@@ -1,7 +1,26 @@
-export 'package:gymnastic_center/presentation/screens/posts/posts_screen.dart';
-export 'package:gymnastic_center/presentation/screens/videos/video_player/video_player_screen.dart';
-export 'package:gymnastic_center/presentation/screens/courses/course_screen.dart';
 export 'package:gymnastic_center/presentation/screens/home_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/account/account_details_screen.dart';
+export 'package:gymnastic_center/presentation/screens/account/account_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/auth/login_screen.dart';
+export 'package:gymnastic_center/presentation/screens/auth/register_screen.dart';
+export 'package:gymnastic_center/presentation/screens/auth/start_screen.dart';
+export 'package:gymnastic_center/presentation/screens/auth/welcome_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/courses/course_screen.dart';
 export 'package:gymnastic_center/presentation/screens/courses/courses_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/posts/posts_screen.dart';
+export 'package:gymnastic_center/presentation/screens/posts/post_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/notifications/notification_status_screen.dart';
+export 'package:gymnastic_center/presentation/screens/notifications/notifications_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/splash/splash_screen.dart';
+
+export 'package:gymnastic_center/presentation/screens/tabs/settings_screen.dart';
+export 'package:gymnastic_center/presentation/screens/tabs/theme_screen.dart';
+
 export 'package:gymnastic_center/presentation/screens/videos/videos_screen.dart';
-export 'package:gymnastic_center/presentation/screens/register/register_screen.dart';
+export 'package:gymnastic_center/presentation/screens/videos/video_player/video_player_screen.dart';
